+++ publisher-sdk/src/androidTest/java/com/criteo/publisher/degraded/StandaloneDegradedTest.java
@@ -107,7 +107,7 @@ public class StandaloneDegradedTest {
   @Test
   public void whenLoadingAnInterstitial_ShouldNotDoAnyCallToCdb() throws Exception {
     runOnMainThreadAndWait(() -> {
-      CriteoInterstitial interstitial = new CriteoInterstitial(context, interstitialAdUnit);
+      CriteoInterstitial interstitial = new CriteoInterstitial(interstitialAdUnit);
 
       interstitial.loadAd();
     });
@@ -122,7 +122,7 @@ public class StandaloneDegradedTest {
     CriteoInterstitialAdListener listener = mock(CriteoInterstitialAdListener.class);
 
     CriteoInterstitial interstitial = callOnMainThreadAndWait(() ->
-        new CriteoInterstitial(context, interstitialAdUnit));
+        new CriteoInterstitial(interstitialAdUnit));
 
     interstitial.setCriteoInterstitialAdListener(listener);